+++ Constants.h
@@ -1,8 +1,8 @@
 namespace Constants
 {
-  const int POT1_PIN =            A0;
-  const int LEFT_BUTTON_PIN =     10;
-  const int RIGHT_BUTTON_PIN =     9;
-  const int UP_BUTTON_PIN =        8;
-  const int DOWN_BUTTON_PIN =      7;
+  const byte POT1_PIN =            A0;
+  const byte LEFT_BUTTON_PIN =     10;
+  const byte RIGHT_BUTTON_PIN =     9;
+  const byte UP_BUTTON_PIN =        8;
+  const byte DOWN_BUTTON_PIN =      7;
 }
